+++ lua/obsidian/lsp/handlers/document_symbol.lua
@@ -2,15 +2,14 @@ local obsidian = require "obsidian"
 
 ---@param _ lsp.DocumentSymbolParams
 return function(_, handler)
-  local note = obsidian.Note.from_buffer(0, { collect_anchor_links = true })
-
-  if not note then -- HACK: somehow DocumentSymbol gets called with no valid note
-    return
-  end
-
   ---@type lsp.DocumentSymbol[]
   local symbols = {}
   local lookup = {}
+
+  local note = obsidian.Note.from_buffer(0, { collect_anchor_links = true })
+  if not note then -- HACK: somehow DocumentSymbol gets called with no valid note
+    return handler(nil, symbols)
+  end
 
   for _, anchor in pairs(note.anchor_links) do
     local display = string.rep("#", anchor.level) .. " " .. anchor.header