+++ tests/test_api.lua
@@ -1,6 +1,5 @@
 local new_set, eq = MiniTest.new_set, MiniTest.expect.equality
 local h = dofile "tests/helpers.lua"
-local Path = require "obsidian.path"
 
 local T, child = h.child_vault {
   pre_case = [[M = require"obsidian.api"]],
@@ -105,57 +104,6 @@ T["cursor_heading"] = function()
   eq(1, child.lua([[return M.cursor_heading()]]).level)
   child.api.nvim_win_set_cursor(0, { 2, 0 })
   eq(vim.NIL, child.lua [[return M.cursor_heading()]])
-end
-
--- TODO: move to lsp/test_definition.lua
-T["follow_link"] = new_set()
-
-local function write(str, path)
-  vim.fn.writefile(vim.split(str, "\n"), tostring(path))
-end
-
-T["follow_link"]["follow wiki links"] = function()
-  local referencer = [==[
-
-[[target]]
-]==]
-
-  local root = Path.new(child.lua_get [[tostring(Obsidian.dir)]])
-  local referencer_path = root / "referencer.md"
-  write(referencer, referencer_path)
-
-  local target_path = root / "target.md"
-  write("", target_path)
-
-  child.lua(string.format([[vim.cmd("edit %s")]], referencer_path))
-  child.api.nvim_win_set_cursor(0, { 2, 0 })
-  local cursor_link = child.lua_get [[M.cursor_link()]]
-  eq("[[target]]", cursor_link)
-  local action = string.format([[M.follow_link("%s")]], cursor_link)
-  child.lua_get(action)
-  eq(tostring(root / "target.md"), child.api.nvim_buf_get_name(0))
-end
-
-T["follow_link"]["follow markdown links"] = function()
-  local referencer = [==[
-
-[target](./target.md)
-]==]
-
-  local root = Path.new(child.lua_get [[tostring(Obsidian.dir)]])
-  local referencer_path = root / "referencer.md"
-  write(referencer, referencer_path)
-
-  local target_path = root / "target.md"
-  write("", target_path)
-
-  child.lua(string.format([[vim.cmd("edit %s")]], referencer_path))
-  child.api.nvim_win_set_cursor(0, { 2, 0 })
-  local cursor_link = child.lua_get [[M.cursor_link()]]
-  eq("[target](./target.md)", cursor_link)
-  local action = string.format([[M.follow_link("%s")]], cursor_link)
-  child.lua_get(action)
-  eq(tostring(root / "target.md"), child.api.nvim_buf_get_name(0))
 end
 
 return T