+++ tests/test_util.lua
@@ -215,6 +215,7 @@ T["parse"]["header"]["should strip white space at the end"] = function()
   eq({ header = "Hello World", level = 2, anchor = "#hello-world" }, M.parse_header "## Hello World ")
 end
 
+-- TODO: test for all link types
 T["parse"]["link"] = new_set()
 
 T["parse"]["link"]["should parse link"] = function()